requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
pytest
httpx
jinja2